Rename server module and remove microstructure references
