+++ docker-compose.override.yml
@@ -1,0 +1,5 @@
+version: '3.9'
+services:
+  web:
+    environment:
+      - APP_DEBUG=true
