+++ pyproject.toml
@@ -1,0 +1,38 @@
+[build-system]
+requires = ["setuptools>=64", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ml_server"
+version = "0.1.0"
+description = "Flask and Celery based microstructural analysis server"
+readme = "README.md"
+requires-python = ">=3.10"
+dependencies = [
+    "Flask==3.0.2",
+    "Werkzeug==3.0.1",
+    "Jinja2==3.1.2",
+    "python-dotenv==1.0.0",
+    "Pillow==11.1.0",
+    "requests==2.32.4",
+    "celery==5.3.6",
+    "redis==5.0.1",
+    "gunicorn==21.2.0",
+    "Flask-Compress==1.14",
+    "Flask-Talisman==1.1.0",
+    "prometheus-client==0.20.0"
+]
+
+[project.optional-dependencies]
+test = [
+    "pytest==7.4.0",
+    "pytest-cov==4.1.0",
+    "numpy>=1.26.0",
+    "flake8",
+    "black",
+    "isort",
+    "mypy"
+]
+
+[project.scripts]
+ml-server = "app:main"
