+++ setup.cfg
@@ -1,0 +1,10 @@
+[metadata]
+name = ml_server
+version = 0.1.0
+
+[options]
+packages = find:
+py_modules = app
+
+[options.package_data]
+* = *.txt, *.md
